refactor: move action types back to store
and leave auth-response.interface in models.
