+++ airways/src/app/core/models/auth-response.interface.ts
@@ -1,0 +1,9 @@
+export interface AuthResponseData {
+  accessToken: string;
+  user: User;
+}
+
+export interface User {
+  id: string;
+  email: string;
+}
